python_fastapi/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
docker==7.0.0
kubernetes==10.0.1
conjur-client==0.1.0
